// types/chat.ts
import { OpenAIModel } from './openai'

export interface Message {
    role: Role
    content: string
}

export type Role = 'assistant' | 'user'

export interface ChatBody {
  model: OpenAIModel
  messages: Message[]
  key: string
}

export interface Conversation {
    id: string
    name: string
    messages: Message[]
    model: OpenAIModel
}